requirements.txt
git+https://github.com/knikolla/coldfront@coldfront_plugin_cloud_urls#egg=coldfront
mozilla-django-oidc
djangorestframework
python-cinderclient  # TODO: Set version for OpenStack Clients
python-keystoneclient
python-memcached==1.59
python-novaclient
python-neutronclient
python-swiftclient
